+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 - Added handling of window resizing (#42)
 - Re-enabled timer widget
 - Fixed mouse movement in a widget based on its screen position (#34)
+- Added window (un)focus event handling (#41)
 
 ## 0.1.18